restructure test organization
